Changes for #9 review - session init with token.
Updated docs for DiscordOAuthSession: added warning and note.

Made 'code' and 'token' both required keyword arguments. This class isn't meant to be initialized by users so it's probably best to force both for clarity.

Added properties for cached discord data and removed use of cache from the request methods, to allow re-requests to be made easier.